requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
argon2-cffi-bindings==21.2.0; python_version >= '3.6'
argon2-cffi==21.3.0
asgiref==3.4.1; python_version >= '3.6'
backports.zoneinfo==0.2.1; python_version < '3.9'
bleach==4.1.0; python_version >= '3.6'
brotli==1.0.9
cffi==1.15.0
dj-database-url==0.5.0
django-cors-headers==3.10.1
django-environ==0.8.1
django-on-heroku==1.1.2
django-summernote==0.8.20.0
djangorestframework==3.13.1
gunicorn==20.1.0
packaging==21.3; python_version >= '3.6'
pillow==9.0.0
psycopg2-binary==2.9.3
pycparser==2.21
pyparsing==3.0.6; python_version >= '3.6'
pytz==2021.3
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
webencodings==0.5.1
whitenoise[brotli]==5.3.0
